Move getPythonVersion().
