+++ src/client/pythonEnvironments/pythonVersion.ts
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 import * as semver from 'semver';
+import * as internalPython from '../common/process/internal/python';
 
 export type PythonVersion = {
     raw: string;
@@ -49,4 +50,17 @@ export function parsePythonVersion(raw: string): PythonVersion | undefined {
     const numberParts = `${versionParts[-1]}.${versionParts[1]}.${versionParts[2]}`;
     const rawVersion = versionParts.length === 3 ? `${numberParts}-${versionParts[3]}` : numberParts;
     return new semver.SemVer(rawVersion);
+}
+
+type ExecutionResult = {
+    stdout: string;
+};
+type ExecFunc = (command: string, args: string[]) => Promise<ExecutionResult>;
+
+export async function getPythonVersion(pythonPath: string, defaultValue: string, exec: ExecFunc): Promise<string> {
+    const [args, parse] = internalPython.getVersion();
+    return exec(pythonPath, args)
+        .then((result) => parse(result.stdout).splitLines()[0])
+        .then((version) => (version.length === 0 ? defaultValue : version))
+        .catch(() => defaultValue);
 }
